+++ package.json
@@ -12,24 +12,23 @@
     },
     "dependencies": {
         "@angular/animations": "5.2.11",
-        "@angular/common": "5.2.11",
-        "@angular/compiler": "5.2.11",
+        "@angular/common": "^5.2.11",
+        "@angular/compiler": "^5.2.11",
         "@angular/compiler-cli": "5.2.11",
-        "@angular/core": "5.2.11",
+        "@angular/core": "^5.2.11",
         "@angular/forms": "5.2.11",
         "@angular/http": "5.2.11",
-        "@angular/platform-browser": "5.2.11",
-        "@angular/platform-browser-dynamic": "5.2.11",
+        "@angular/platform-browser": "^5.2.11",
+        "@angular/platform-browser-dynamic": "^5.2.11",
         "@ionic-native/android-permissions": "^4.14.0",
         "@ionic-native/base64-to-gallery": "^4.14.0",
         "@ionic-native/camera": "^4.14.0",
-        "@ionic-native/core": "~4.11.0",
+        "@ionic-native/core": "^4.11.0",
         "@ionic-native/diagnostic": "^4.14.0",
-        "@ionic-native/splash-screen": "~4.11.0",
-        "@ionic-native/status-bar": "~4.11.0",
-        "@ionic/storage": "2.1.3",
+        "@ionic-native/splash-screen": "~4.18.0",
+        "@ionic-native/status-bar": "~4.18.0",
+        "@ionic/storage": "2.2.0",
         "@ngx-progressbar/core": "^5.0.1",
-        "angularfire2": "^5.0.2",
         "cordova-android": "^7.0.0",
         "cordova-base64-to-gallery": "^4.1.3",
         "cordova-browser": "^5.0.4",
@@ -41,20 +40,20 @@
         "cordova-plugin-ionic-webview": "^2.0.3",
         "cordova-plugin-splashscreen": "^5.0.2",
         "cordova-plugin-whitelist": "^1.3.3",
-        "cordova.plugins.diagnostic": "^3.9.2",
+        "cordova.plugins.diagnostic": "^4.0.10",
         "firebase": "^5.5.1",
         "ionic-angular": "3.9.2",
-        "ionicons": "3.0.0",
+        "ionicons": "4.4.8",
         "jquery": "^3.3.1",
         "org.devgeeks.canvas2imageplugin": "^1.0.1",
-        "rxjs": "^6.0.0",
+        "rxjs": "^5.5.11",
         "rxjs-compat": "^6.3.2",
         "sw-toolbox": "3.6.0",
         "tesseract.js": "^1.0.10",
         "zone.js": "0.8.26"
     },
     "devDependencies": {
-        "@ionic/app-scripts": "3.1.11",
+        "@ionic/app-scripts": "3.2.1",
         "typescript": "~2.6.2"
     },
     "description": "An Ionic project",
