+++ src/pages/process/process.html
@@ -1,0 +1,37 @@
+<!--
+  Generated template for the ProcessPage page.
+
+  See http://ionicframework.com/docs/components/#navigation for more info on
+  Ionic pages and navigation.
+-->
+<ion-header>
+  <ion-navbar>
+    <button ion-button menuToggle>
+      <ion-icon name="menu"></ion-icon>
+    </button>
+    <ion-title>OCR <small><code>(* {{subTitle ? subTitle : 'default' }})</code></small></ion-title>
+  </ion-navbar>
+</ion-header>
+<ion-content style="overflow: auto">
+  <label for="getImg">
+    <img ion-row width="100%" *ngIf="!selectedImage" class="" src="../../assets/imgs/img-default.png">
+    <img *ngIf="selectedImage" src="{{selectedImage}}" alt="">
+
+  </label>
+  <!-- <img src="{{base64Data}}" class="img-responsive" alt="Image"> -->
+  <canvas #layout id="canvasInput" (touchstart)="start($event)" style="display: none"></canvas>
+  <div class="canvas"></div>
+  <!-- <canvas #output id="canvasOutput" style="margin: 0 auto; display: none"></canvas> -->
+  <div>
+    <button type="button" id="getImg" ion-button (click)="getPicture(indexOfTemplate)" style="display: none">Library</button>
+  </div>
+</ion-content>
+<ion-footer>
+  <ion-toolbar style="text-align: center">
+    <button round ion-button (click)="takePicture(indexOfTemplate)">
+      <ion-icon name="camera"></ion-icon>
+    </button>
+    <button type="button" round ion-button (click)="processImage(indexOfTemplate)" [disabled]='!selectedImage'>Process</button>
+
+  </ion-toolbar>
+</ion-footer>
