+++ src/pages/process/process.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { IonicPageModule } from 'ionic-angular';
+import { ProcessPage } from './process';
+
+@NgModule({
+  declarations: [
+    ProcessPage,
+  ],
+  imports: [
+    IonicPageModule.forChild(ProcessPage),
+  ],
+})
+export class ProcessPageModule {}
