+++ src/pages/process/process.scss
@@ -1,0 +1,7 @@
+page-process {
+
+    label{
+        display: block;
+        margin-top:20%;
+    }
+}
