+++ src/providers/share-data/share-data.ts
@@ -1,0 +1,17 @@
+import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+
+/*
+  Generated class for the ShareDataProvider provider.
+
+  See https://angular.io/guide/dependency-injection for more info on providers
+  and Angular DI.
+*/
+@Injectable()
+export class ShareDataProvider {
+
+  constructor(public http: HttpClient) {
+    console.log('Hello ShareDataProvider Provider');
+  }
+
+}
